Move libgit2 settings out of util
